episode page refactor
